Implement rendering of thermal grids to a pixel map
A copy of the DejaVu Sans font is embedded in the binary now. The font
has been minimized so that it only includes the glyphs needed to render
temperatures:

+-0123456789.,

A few extra characters are included for now such as various narrow
spaces to leave open the possibility of localized formatting.