Implement main.rs as a simple MJPEG stream over HTTP
There's no actual configuration yet, and it doesn't actually do any
occupancy determinations, but it's a start. It'll do 10FPS with ~30% of
one core on a Raspberry Pi 4B. It'll max out a BeagleBone and drop
frames, but the image isn't corrupted (like room-assistant would do when
it pegged the CPU).