Expand Camera config implementation
From is now implemented for CameraSettings to provide easier access to
the config structures inside. The frame rate is no longer optional,
being set by the individual camera implementations (as they know what
each camera is capable of). From is also implemented for Duration to
make it easier to set that up.

Minimal set up in the main executable to use the config parsing code has
been added as well.