Add a broadcasting sink with non-lagging streams
Mostly just a wrapper around `tokio_stream::wrapper::BroadcastStream`
and an implementation of `futures::Sink` for
`tokio::sync::broadcast::Sender`. The intention is to use this to
connect the camera frame output to both the rendering and occupancy
detection functions, and to connect the rendering output to the
different streaming implementations.