Always use the image crate
No longer an optional feature, as more of it is going to be used
throughout the program. The JPEG encoder *could* be made optional, but
it's simpler to just include it. I'm considering ripping out mozjpeg,
which will make the cross builds even simpler as I'm not sure it makes
that much of a difference performance-wise.