Add a moving average implementation
Not used for anything right now, but I've been playing with adding a
filter to the camera frames to try to smooth out some noise. It does
smoth it out, a little, but it's not what I was looking for. I'm
thinking of maybe instead putting a filter on the output of the pixel
classifier instead, to cut out pixels on the edge of being considered
"active". In any case, I like the code and want to keep it around for a
while.