Refacto spmc under pubsub module and add count_tree
TreeCount is kinda like a hierarchal semaphore. You can create a tree of
the counters, and create CountTokens from them which increments the
value of a tree node as well as any parent nodes. TreeCounts are Futures,
so you can await on them. The condition is for the count to go above 0,
and the yielded value is the total number of outstanding CountTokens.
When CountTokens are dropped, the value of the source node (as well as
any parents) is decremented.