First pass at implementing Tracker
Right now it just uses a static threshold, and then counts the number of
different connected regions in the image. There's lots of room for
improvement, like looking at the current ambient temperature and
comparing the blob temperatures with that, or doing actual motion
estimation, etc. This is a first pass at something the kind of works.