Rewrite mqtt::State
State is now an enum that covers basic usage (just publish MQTT messages
as the state changes) and with Home Assistant (publish a persistent
config message for HAss to create entities from, and update an
availability topic for the entire device).

The plan is to remove mqtt::client, as most of the logic from there is
now in State. A longer-term goal is to implement Sink on State, as I was
having trouble implementing it and I'm not confident enough in my
understanding of Pin and assorted async internals to implement it
currently.