Rewrite camera configuration handling
There's still a flattened enum, but now it's just of the CameraKind, and
the I2C settings for that camera (leaving it open for non-I2C cameras in
the future). The frame rate validation is now done later on, when the
smahed camera is being constructed, so those tests have been moved for
now. That should also address a problem where non-native frame rates
were being rejected for GridEYE cameras.