Rewrite how SharedCamera captures measurements and times frames
Planning on adding an actual PID implementation later, but right now
this is much better than the previous implementation. It adapts to
cameras running a little off, and can also batch operations together
(ex: Melexis cameras measureing the temperature while capturing the
image).

The ThermalCamera implementations are no longer generic on the I2C
implementation now; this might be added back in the future if it makes
testing easier, but this makes the code a bit nicer to read.