Rewrite renderer
This is set up for hopefully implementing GPU-based rendering later.

Rendering now happens on different layers, with the background layer
consisting of the colored grid squares and the temperature text as a
mask on top of that. The text color selection happens when the layers
are being flattened, allowing temperatures to be either upscaled for the
bacground, or downscaled for the text.

Dynamic temperature scale limits are also much improved, now being based
on a moving average of the last ten frames, and enforcing a minimum
range between the upper and lower limits. This results in a much nicer
display by default.