Rewrite camera settings handling
No more manually implemented Deserialize, and this is the first chunk of
a larger rewrite of how CLI arguments are merged in with config file
settings.

The main reason this was done was to clean up how the conditional
compilation for mock_camera is handled. The blocks that handle
mock-camera specific functions are self contained and it's easier to see
what they're doing.

The process for creating ThermalCamera instances is also more
self-contained, with the camera kind specific code also contained with
camera/settings.rs.