Add object tracker tests with real-world data
I've added four recorded scenes covering a couple different scenarios
along with a description of each scene and the frame numbers at which
the occupancy count should change. Currently most of these tests fail,
and the walk-in test is non-deterministic (and I don't know why...).

If future scenes are added, they should be included as the raw bincode
files. TOML cannot represent the RecordedData structure properly (no
root-level arrays), and both plain-text formats (YAML and TOML) are
*much* larger than the bincode. For example, `walk-in.bin` is 7.5MB, but
the pretty-printed `walk-in.yml` is 24MB. git compresses data using
deflate, but even then, the file sizes are in favor of bincode (5.1MB vs
8.3MB).